Add scala-native-p, a NIR deserialization program
It is based on the dumping feature of Scala Native, but slighly
redesigned to allow directly passing regular files and printing to
stdout instead of saving to a file. Because of that, it requires the
Scala Native utils dependency.

It features --help, --version and --classpath options. The last one allows to
pass a classpath where the NIR files are meant to be found. If not
passed, the files are looked for in cwd.