Add initial `scala-native-p` script (#7)
* Add scala-native-p, a NIR deserialization program

It is based on the dumping feature of Scala Native, but slighly
redesigned to allow directly passing regular files and printing to
stdout instead of saving to a file. Because of that, it requires the
Scala Native utils dependency.

It features --help, --version and --classpath options. The last one allows to
pass a classpath where the NIR files are meant to be found. If not
passed, the files are looked for in cwd.

* Update sbt build to use a default main class

* Set main class in assembly

* Use NIR classloader to read contents of the files

* Add bash/batch scripts

* Adjust standalone scripts to allow testing scala-native-p

* Add mising blacnk lines in scripts

* Run scalafmt

Co-authored-by: Jan Chyb <jan.chyb.99@gmail.com>